pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "workflows*"]
exclude = ["runs*", "experiment_logs*", "experiment_results*", "tests*", "visualization*", "langsmith_traces*"]

[project]
name = "agentic-eng-design"
version = "1.0.0"
description = "Multi-agent LLM framework for conceptual systems engineering and design"
authors = [
    {name = "Soheyl Massoudi", email = "smassoudi@umd.edu"},
    {name = "Mark Fuge", email = "fuge@umd.edu"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # Core LangChain and LangGraph packages
    "langgraph>=0.2.72",
    "langchain>=0.3.18",
    "langchain-core>=0.3.35",
    "langchain-community>=0.3.17",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.5",
    "langchain-ollama>=0.2.3",
    # Data processing and scientific computing
    "pydantic>=2.10.6",
    "numpy>=1.26.4",
    "scipy>=1.15.2",
    "pandas>=2.2.3",
    "networkx>=3.4.2",
    # Visualization
    "plotly>=6.0.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    # LLM providers
    "openai>=1.61.0",
    "ollama>=0.4.7",
    # Search and information retrieval tools
    "arxiv>=2.1.3",
    "ddgs>=7.5.1",
    "wikipedia>=1.4.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "ipython>=8.27.0",
    "requests>=2.32.3",
    "langsmith>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.10.0",
    "pre-commit>=3.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/SoheylM/agentic-eng-design"
Paper = "https://doi.org/10.1115/1.4070328"

# ──────────────────────────────────────────────
#  Ruff – linting and formatting
# ──────────────────────────────────────────────
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "runs/",
    "experiment_logs/",
    "experiment_results/",
    "langsmith_traces/",
    "*.ipynb",
    "*.md",
    "*.json",
]

line-length = 120  # longer lines acceptable for LLM prompt strings
target-version = "py311"

[tool.ruff.lint]
select = [
    "E4",   # pycodestyle errors (import)
    "E7",   # pycodestyle errors (statement)
    "E9",   # pycodestyle errors (runtime)
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "PIE",  # flake8-pie
    "PTH",  # flake8-use-pathlib
    "PGH",  # pygrep-hooks
    "RUF",  # Ruff-specific rules
]

ignore = [
    "E501",    # line too long – handled by formatter, and prompt strings are long
    "B008",    # do not perform function calls in argument defaults
    "C901",    # too complex
    "SIM108",  # use ternary operator – sometimes less readable
    "UP007",   # use X | Y for union types – requires 3.10+ runtime
]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
known-first-party = ["agents", "workflows"]

[tool.ruff.lint.per-file-ignores]
# Allow print statements in agent nodes (they log progress)
"agents/*.py" = ["T20"]
"workflows/*.py" = ["T20"]
"run_pipeline.py" = ["T20"]
"eval_saved.py" = ["T20"]
"eval_all.py" = ["T20"]
"display_metrics.py" = ["T20"]
# Tests can use magic values and assertions
"tests/**/*" = ["PLR2004", "S101"]

# ──────────────────────────────────────────────
#  MyPy – type checking
# ──────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
files = ["*.py", "agents/", "workflows/"]
exclude = [
    "^runs/",
    "^experiment_logs/",
    "^experiment_results/",
    "^langsmith_traces/",
    "^tests/",
    "^visualization/",
    "^build/",
]

[[tool.mypy.overrides]]
module = [
    "langchain_core.*",
    "langchain_openai.*",
    "langchain_ollama.*",
    "langchain_community.*",
    "langchain_experimental.*",
    "langgraph.*",
    "plotly.*",
    "networkx.*",
    "IPython.*",
    "arxiv.*",
]
ignore_missing_imports = true

# ──────────────────────────────────────────────
#  Pytest
# ──────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
